add odomcomponent class to separate from Movementcomponent
