+++ Source/RapyutaSimulationPlugins/Private/Drives/RobotVehicleMovementComponent.cpp
@@ -20,9 +20,6 @@
 void URobotVehicleMovementComponent::Initialize()
 {
     OwnerVehicle = CastChecked<ARRBaseRobot>(GetOwner());
-    GaussianRNGPosition = std::normal_distribution<>{NoiseMeanPos, NoiseVariancePos};
-    GaussianRNGRotation = std::normal_distribution<>{NoiseMeanRot, NoiseVarianceRot};
-
     InitData();
 }
 
@@ -66,17 +63,17 @@ void URobotVehicleMovementComponent::TickComponent(float InDeltaTime,
         else
         {
             Super::TickComponent(InDeltaTime, TickType, ThisTickFunction);
-
+            UE_LOG_WITH_INFO(LogRapyutaCore, Warning, TEXT("TickComponent non AI "));
             // Make sure that everything is still valid, and that we are allowed to move.
             if (IsValid(UpdatedComponent))
             {
+                UE_LOG_WITH_INFO(LogRapyutaCore, Warning, TEXT("TickComponent non AI2 "));
                 //1- Update vels to OwnerVehicle's target vels
                 Velocity = OwnerVehicle->TargetLinearVel;
                 AngularVelocity = OwnerVehicle->TargetAngularVel;
 
                 //2- Movement control for OwnerVehicle
                 UpdateMovement(InDeltaTime);
-                UpdateOdom(InDeltaTime);
 
                 //3- Update OwnerVehicle's velocity to [Velocity], must be after [UpdateMovement()]
                 UpdateComponentVelocity();
@@ -236,105 +233,9 @@ void URobotVehicleMovementComponent::UpdateMovement(float InDeltaTime)
     }
 }
 
-void URobotVehicleMovementComponent::SetFrameIds(const FString& InFrameId, const FString& InChildFrameId)
-{
-    OdomData.Header.FrameId = FrameId = InFrameId;
-    OdomData.ChildFrameId = ChildFrameId = InChildFrameId;
-}
-
-// todo separate ROS
-void URobotVehicleMovementComponent::InitOdom()
-{
-    AActor* owner = GetOwner();
-    OdomData.Header.FrameId = FrameId;
-    OdomData.ChildFrameId = ChildFrameId;
-
-    if (OdomSource == EOdomSource::ENCODER)
-    {    // odom source = encoder. Odom frame start from robot initial pose
-        InitialTransform.SetTranslation(owner->GetActorLocation());
-        InitialTransform.SetRotation(owner->GetActorQuat());
-    }
-    else
-    {    // odom source = world. Odom frame start from world origin
-        InitialTransform.SetTranslation(FVector::ZeroVector);
-        InitialTransform.SetRotation(FQuat::Identity);
-    }
-
-    OdomData.Pose.Pose.Position = InitialTransform.GetTranslation();
-    OdomData.Pose.Pose.Orientation = InitialTransform.GetRotation();
-
-    PreviousTransform = InitialTransform;
-    PreviousNoisyTransform = InitialTransform;
-
-    // todo temporary hardcoded
-    OdomData.Pose.Covariance[0] = 1e-05f;
-    OdomData.Pose.Covariance[7] = 1e-05f;
-    OdomData.Pose.Covariance[14] = 1e+12;
-    OdomData.Pose.Covariance[21] = 1e+12;
-    OdomData.Pose.Covariance[28] = 1e+12;
-    OdomData.Pose.Covariance[35] = 1e-03f;
-
-    OdomData.Twist.Covariance[0] = 1e-05f;
-    OdomData.Twist.Covariance[7] = 1e-05f;
-    OdomData.Twist.Covariance[14] = 1e+12;
-    OdomData.Twist.Covariance[21] = 1e+12;
-    OdomData.Twist.Covariance[28] = 1e+12;
-    OdomData.Twist.Covariance[35] = 1e-03f;
-
-    bIsOdomInitialized = true;
-}
-
-void URobotVehicleMovementComponent::UpdateOdom(float InDeltaTime)
-{
-    if (!bIsOdomInitialized)
-    {
-        InitOdom();
-    }
-
-    // time
-    OdomData.Header.Stamp = URRConversionUtils::FloatToROSStamp(UGameplayStatics::GetTimeSeconds(GetWorld()));
-
-    // previous estimated data (with noise)
-    FVector previousEstimatedPos = PreviousNoisyTransform.GetTranslation();
-    FQuat previousEstimatedRot = PreviousNoisyTransform.GetRotation();
-
-    AActor* owner = GetOwner();
-
-    // position
-    FVector pos = InitialTransform.GetRotation().UnrotateVector(owner->GetActorLocation() - InitialTransform.GetTranslation());
-    FVector previousPos = PreviousTransform.GetTranslation();    // prev pos without noise
-    PreviousTransform.SetTranslation(pos);
-    pos += previousEstimatedPos - previousPos + WithNoise * FVector(GaussianRNGPosition(Gen), GaussianRNGPosition(Gen), 0);
-
-    FRotator noiseRot = FRotator(0, 0, WithNoise * GaussianRNGRotation(Gen));
-    FQuat rot = owner->GetActorQuat() * InitialTransform.GetRotation().Inverse();
-    FQuat previousRot = PreviousTransform.GetRotation();
-    PreviousTransform.SetRotation(rot);
-    rot = noiseRot.Quaternion() * previousEstimatedRot * previousRot.Inverse() * rot;
-    rot.Normalize();
-
-    PreviousNoisyTransform.SetTranslation(pos);
-    PreviousNoisyTransform.SetRotation(rot);
-
-    OdomData.Pose.Pose.Position = pos + RootOffset.GetTranslation();
-    OdomData.Pose.Pose.Orientation = rot;
-
-    OdomData.Twist.Twist.Linear = OdomData.Pose.Pose.Orientation.UnrotateVector(pos - previousEstimatedPos) / InDeltaTime;
-    OdomData.Twist.Twist.Angular =
-        FMath::DegreesToRadians((rot * previousEstimatedRot.Inverse()).GetNormalized().Euler()) / InDeltaTime;
-
-    OdomData.Pose.Pose.Orientation *= RootOffset.GetRotation();
-}
-
-FTransform URobotVehicleMovementComponent::GetOdomTF() const
-{
-    return FTransform(OdomData.Pose.Pose.Orientation, OdomData.Pose.Pose.Position);
-}
 
 void URobotVehicleMovementComponent::InitData()
 {
-    InitOdom();
-
     AActor* owner = GetOwner();
     ContactPoints.Empty();
     TArray<UActorComponent*> actorContactPoints = owner->GetComponentsByTag(USceneComponent::StaticClass(), TEXT("ContactPoint"));
