+++ Source/RapyutaSimulationPlugins/Public/Tools/RRROS2OdomPublisher.h
@@ -15,6 +15,7 @@
 
 // RapyutaSimulationPlugins
 #include "Tools/RRROS2TFPublisher.h"
+#include "Tools/RRROS2BaseSensorPublisher.h"
 
 #include "RRROS2OdomPublisher.generated.h"
 
@@ -26,7 +27,7 @@ class ARRBaseRobot;
  * @sa [UROS2Publisher](https://rclue.readthedocs.io/en/devel/doxygen_generated/html/d6/dd4/class_u_r_o_s2_publisher.html)
  */
 UCLASS(ClassGroup = (Custom), Blueprintable, meta = (BlueprintSpawnableComponent))
-class RAPYUTASIMULATIONPLUGINS_API URRROS2OdomPublisher : public UROS2Publisher
+class RAPYUTASIMULATIONPLUGINS_API URRROS2OdomPublisher : public URRROS2BaseSensorPublisher
 {
     GENERATED_BODY()
 
@@ -37,25 +38,12 @@ public:
      */
     URRROS2OdomPublisher();
 
-    UPROPERTY(BlueprintReadWrite)
-    TWeakObjectPtr<ARRBaseRobot> RobotVehicle = nullptr;
+    virtual bool InitializeWithROS2(UROS2NodeComponent* InROS2Node) override;
 
     UPROPERTY(BlueprintReadWrite)
     URRROS2TFPublisher* TFPublisher = nullptr;
 
     void InitializeTFWithROS2(UROS2NodeComponent* InROS2Node);
-
-    bool InitializeWithROS2(UROS2NodeComponent* InROS2Node) override;
-
-    /**
-     * @brief Initialize odom publisher
-     * @todo is this method necessary?
-     */
-    UFUNCTION(BlueprintCallable)
-    void InitOdomPublisher(UROS2NodeComponent* InROS2Node)
-    {
-        InitializeWithROS2(InROS2Node);
-    }
 
     // void RevokeUpdateCallback() override;
     void UpdateMessage(UROS2GenericMsg* InMessage) override;
